LayoutTests/http/tests/lazyload/synchronous-frame-creation-expected.txt

Tests that a lazy loaded iframe gets a contentWindow / contentDocument as soon as it gets inserted in the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


document.body.append(iframe)
PASS iframe.contentDocument.URL is "about:blank"
PASS iframe.contentWindow.location.href is "about:blank"
PASS successfullyParsed is true

TEST COMPLETE
